Preparing for more factory classes.
